refactor(goods): SelectedOption type을 shared/types로 분리
