+++ src/shared/types/selectedOption.ts
@@ -1,0 +1,6 @@
+import { GoodDetailOption } from "@/entities/goods/goods.types";
+
+export type SelectedOption = {
+  type: "option" | "extraOption";
+  quantity: number;
+} & GoodDetailOption;
